Fill up the number of players in the frontend to reach a defined number of minimum players
This fixes a bug, where there are not enough player forms in the Frontend when the Backend sends a list of players already added to the game. The current seeder of the Backend adds only one dummy player to a game which means there was only one Player form in the Frontend as well.